apps/reciepe_agent/pyproject.toml
[project]
name = "reciepe-agent"
version = "0.1.0"
description = "A LangGraph-based recipe agent"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "langchain-google-genai>=0.3.0"
]

[tool.poe.tasks]
serve = "uvicorn main:app --reload --host 0.0.0.0 --port 8000"
dev = "python main.py"
test = "python -m pytest tests/"
graph = "langgraph dev"
inspect = "langgraph inspect"

[tool.poe.tasks.agent]
cmd = "python main.py"
help = "Run the Recipe Agent"
